face-recognition-service/requirements.txt
flask>=2.3.0,<4.0.0
face_recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
